Update advanced_agent.py
If the agent reaches desperation 10 or higher, it will attempt to kick down locked doors to open up new areas. (It currently doesn't actually check which doors actually lead to new areas, though.)
The agent will now mark traps on the map and avoid them. If the first thing it sees on a new square is a corpse, it'll assume that it's a trap. (Note: This can occasionally cause a panic if the agent steps off a trap it triggered and into a corridor with nothing in it, because it would rather panic than step on the trap. This behavior will be fixed at some point in the future, but it's not a high priority issue right now.)