Enable picking up items
Added a basic framework for recognizing items the agent stands on. Right now, it's really primitive, but this framework lays the foundation for the agent to interact with items much more intelligently in the future. For now, the biggest improvement is that the agent will pick up basic food (not corpses) it steps on, which is a boon for classes that have little or no food in their starting inventory.

Note that currently, the agent can't really detect when it's standing on a pile with more than one item, because then the message appears in a "pop-up", and the NLE currently doesn't make the information in pop-up windows very accessible. (I could in theory grab the message from the screen, but that's sooooooo annoying, and the code for that would end up really obtuse and hard to modify. I'd really rather not.)