Move towards items
Agent will now actively pursue items it sees on the map, rather than being limited to just picking up what it coincidentally happens to step on. Also: Squashed a bug that caused a rare index-out-of-bounds exception if the agent found itself on the bottom edge of the map while desperate enough to consider kicking down doors.