Enabled blind eye-killing
Some classes have a small chance of starting with a blindfold or towel in their inventory. Blinding yourself with one of these items can be used to kill a Floating Eye without incurring its very dangerous gaze attack. This strategy will now be employed, which should slightly reduce the odds of panicking due to being stuck between an eye and a hard place.

I also resolved a couple more edge-of-screen index-out-of-bounds exceptions.