Enabled combat at range
Agent will now use basic throwable weapons. Agent will look in straight lines around itself to look for targets, and if there are none, it sees if it can move to a cardinally adjacent tile to line up with an enemy. Agent now marks pets on the map so it doesn't try to throw a dagger at a monster when a pet is in the way. Note: Launchers are not yet supported; the agent won't yet try to attack with arrows, crossbow bolts, or rocks.