Added iterableOverVicinity
Several times in my code for this agent, I need to check the eight (or sometimes just the cardinal four) squares around me for some purpose. If the agent's position is (x,y), I need to check (x-1,y), (x+1,y), (x,y-1), (x,y+1), (x-1,y-1), (x-1,y+1), (x+1,y-1), and (x+1,y+1) – and the actions corresponding to those eight cases are 0, 2, 3, 1, 7, 4, 6, and 5 respectively. To this point, I've just copy-pasted code eight times and hard-coded the eight versions to handle each direction, but this makes for abhorrent looking code that's really susceptible to all manner of copy-paste errors. To solve this, I've now made iterableOverVicinity, which returns an array consisting of those eight coordinate pairs (sorted so that element n in the array corresponds to action number n), making it MUCH, much easier to handle this kind of situation with a simple for loop. This "golfs" my implementation considerably, and will make it a LOT easier to add all manner of new functionality in the future.

Other new features with this commit include:
-> More narration.
-> Changed how searched_map (the map that tracks how many times we've searched each square in the dungeon) is printed out in the event of a panic. Previously we just outputted the numbers for each square separated by spaces and newlines, but that resulted in an output where numbers weren't aligned to any sort of grid, and rows of output spilled over to more than one line of text in the console, making it practically unreadable. Instead, we now print a single character per square – " " if we haven't searched the square at all, "s" if we've searched it a little bit, and "S" if we've searched it to the point of exhaustion. This produces a grid that's easy to compare/contrast with the main dungeon map.