200x message panic + artificial glyph IDs
Starting with this commit, the agent will panic if it sees the same message two hundred times consecutively. Since most messages display for one step and then disappear, if a message appears THIS many times in a row, it's probably because the agent is repeatedly doing a thing that results in that message. And in turn, that probably means the agent is trying to do something it can't.

The reason we need to panic in this situation is because not all attempted impossible actions cause an in-game turn to pass; and if not, then the agent will most likely just attempt to do that impossible thing over and over again ad infinitum without anything interrupting it. If we didn't panic, the program would run for ages until eventually the agent is terminated by the 10000-steps-without-reward rule in the NLE. That slows down testing a lot.

In addition, we introduce what I call "artificial glyph IDs" in items.py. Some items in Nethack have randomized appearances, such that at the beginning of the game, there's no way to tell if a "forked wand" is a wand of death or wishing or create monster. In some situations, we need to look for a specific item in our inventory, like a potion of acid if we're turning to stone. The NLE's glyphs are inadequate for this purpose because they correspond to the randomized appearances, not the actual identified items, even if the items are in fact formally identified.

So far we do nothing with these artificial glyph IDs, but this is the first baby step towards item identifications. Stay tuned for more on that front.