Item identification added!
Oh, how long I've waited for this day. WE HAVE ITEM IDENTIFICATION!

I mean, it's not anything sophisticated. Basically all we do right now is keep an eye on the descriptions of our inventory and report when we inadvertently formally identify something. Right now, that means either it started in our inventory (those are pre-identified), or we saw a monster using it. But still! We are now able to use healing potions in our starting inventory to survive bad luck early on more consistently. Currently we use healing potions at 25% health or lower; I'll probably increase this HP threshold later, because it's pretty easy for the agent to go from >25% to zero in one hit. But it's still definitely a big help in the agent's adventures, and I can now start implementing usage of scrolls, potions, rings, amulets, and wands in whatever order I see fit. And when we're ready to go farther and start taking active measures to identify items, the backend is there to make it possible. Woohoo!