Cured agent's dendrophobia
Prior to this commit, agent would on rare occasions panic because it was trying to walk through a tree and couldn't. That's now fixed – trees are now correctly treated as impassible.

New narration feature: Narration reports when the agent is afflicted with a status condition such as stunning or blindness, and also reports when the condition is cured or wears off or is otherwise removed. (This functionality has brought to light an issue with the agent's food-eating behavior – if there is a corpse or something underfoot when the agent goes to eat something from its inventory, it'll usually end up eating the food on the ground. That's bad, because that leads to food poisoning. I know what causes it and will fix it when I can.)

Agent is less likely to panic in the face of floating eyes. If the agent has a blindfold and the eye is blocking something the agent wants, the agent will route a path through the floating eye. If the agent is on the verge of panicking, it'll attack a floating eye (not as dangerous as you might think – if the agent is altogether cornered by floating eyes, then those eyes will most likely block any monsters from approaching the agent in its paralyzed state.