Implemented eating corpses
Another major feature! The agent will now eat appropriate corpses off the floor, which partly solves the starving problem non-monk roles had. Not entirely, though... edible corpses aren't too abundant, and our agent errs heavily on the side of caution with respect to the risk of fatal food poisoning and other undesirable effects, so our agent still starves some of the time when not playing monk. More work to be done, I guess. The good news is that the agent now averages 700-1000 score when playing barbarian, which isn't too far behind what monk achieves.

It's worth noting that our present inability to figure out what race we are is now officially an appreciable obstacle to our progress, because we have to pass up every gnome/elf/human corpse we see lest we commit cannibalism.

Other changes with this commit include:
-> Added functionality to report when the agent detects that it kills a monster (currently, only in a way that makes the agent willing to eat a corpse it sees – which means melee combat only, and that zombies aren't reported). I made this message choose randomly between a few different sentences of the form "<verb>ed a(n) <monster>" because I was bored, and also to avoid having my eyes glaze over the narration.
-> Deleted function updateSearchedMapSquare in gamestate.py, which I've labelled as deprecated for some time, because it fundamentally works wrong. I've long since put in a better system for the same purpose anyway.