Pray for nutrition
As a supplement to foodstuffs the agent finds, the agent will Pray to the hero's god in hopes that the god will make everything better. The agent's hunger problems are now about 80% resolved. Prayer isn't entirely reliable (waiting to be 100% certain that the prayer timeout is low enough takes way too long) but the benefits are more than worth the trouble.