Upgrade Skills
Agent will now keep an eye on its Skill level for fighting and weapons, and Enhance those skills when it has the chance. This should give it more of a fighting chance in later floors (it takes a while to use a thing enough for its skill to be eligible for Enhancement, so this isn't really any help for the early floors). The total amount of Enhancing you can do is finite and tied to your LOVE, so we won't be able to enhance every possible skill; but getting enough practice to upgrade our Skills is the limiting factor right now, so we just enhance whatever we can whenever we can.

Other changes with this commit include:
-> Put in some backend for spellcasting. It's currently unused because it's not seaworthy yet, but that's the next major feature I'll be working on.
-> Agent is a little less eager to fight sessile monsters (monsters that don't move and only have passive (reactive) attacks)
-> The @ symbol is no longer marked on the map, because the agent really doesn't need it labelled and it kind of gets in the way of other important labels. (We have the heroRow and heroCol observations for that purpose.) Instead, the map-printing functions check the hero's position and insert the @ symbol onto the output that gets printed to console.
-> As part of making skill tracking work, the agent now checks what class it is (from the ttychars output of the rank title at the beginning of the game). Checking race and alignment is still waiting on that magic menu_strs observation that I'm told will happen one of these days...
-> Re-reworked printing the search map to make it easier to compare against the main dungeon map. (Specifically, floors and non-wall obstacles are represented with different symbols.)
-> Agent is now able to make use of pickaxes. Right now, the agent doesn't try to dig through walls, because undiggable walls are an issue and busting down every wall in sight is hardly the most efficient way of traversing a dungeon floor anyhow; but the agent does use pickaxes to pulverize boulders that might be in the way. Also, if the agent gets desperate while in possession of a pickaxe, it'll dig straight down to go directly to the next floor without needing to find the stairs at all.