Introduced observant_agent
Here is my very first version of my new agent, which I shall name observant_agent. This agent is named for the fact that it's going to make much better use of some of the actions in the NLE that don't have any effect on the game world but change how the agent sees it, which in some cases can give us useful information that isn't shown in the observation otherwise. A couple examples that are up and running right now:
-> Agent uses the interface to name any monster it sees and reserves some memory to keep track of the monster's status. (The actual status-tracking isn't really implemented yet, but the setting aside of memory is. Each individual monster has an object, and I can add functionality to that object as the need arises.)
-> Agent plans out its pathfinding with the help of the extended command #terrain, which hides monsters and objects to show the raw dungeon floor. This FINALLY empowers the agent to move diagonally, since we can be confident we won't try to move diagonally in or out of a square with an open door (the game doesn't allow that). It also makes sure we can spot the stairs even if a monster or item is on top of it.

The other big new feature about this bot is that it runs a LOT faster. It doesn't need to redraw its route every single turn, and it doesn't need to copy over the entire screen of glyphs into its notes every single step. I can't say precisely how much faster it is until I reimplement searching for secret doors, but it's several times faster than organized_agent ever could manage.

Right now, only the barest functionality is implemented for this bot; all it does is name monsters and try to move to unobserved areas or the stairs. There's still a loooooot of functionality left to bring over from organized_agent. I'm committing this, though, because I think this first step was the hardest step. The rest will take time, but it won't be nearly as hard as the parts implemented in this commit.