Fixed --more-- handling
The way my agent handled --more-- prompts borked the functionality of, in particular, monster naming. (Turns out the screen where you're shown all the monster names is a --more-- prompt, which the message handler closed without giving the monster tracker a chance to look. Whoops.) I fixed this by giving monster naming priority over closing out --more-- prompts. BUT, if there's already a --more-- prompt at the beginning of the turn (which would prevent us from pulling up the monster menu), the monster tracker will do nothing so as to give the prompt handler a chance to close out that prompt.

Well, I thought it was a clever and elegant solution, at least.