Searching for secret doors
We now once again search for secret doors. Our new secret door policy is to search the current square a bunch and then look for a suitable wall within a few squares to switch to. It's not perfect (one issue is that theoretically we can get stuck searching a couple of walls with the other dungeon walls too far out to be recognized), but it is good enough for now, and it is fast. All in all, observant_agent is roughly ten times faster than organized_agent. That's no small potatoes.

Also, we don't try to name shopkeepers or invisible monsters. It doesn't really work, and it just drains our limited reserve of names. Other monsters that can't be named will be treated similarly as we encounter them.