Lots of Stuff
This is a bit embarrassing, but... I haven't really been keeping track of all the new features I'm adding with this particular commit. I'll just name a few that come to my mind easily.

* Shopkeepers are now tracked properly, which is an achievement because you can't assign them names – you have to go by their preset name.
* Each monster has its hostile/peaceful/tame status tracked. We make a point of not moving into peaceful monsters to avoid angering them.
* Each monster has its glyph tracked so we know what it is.
* Fixed some issues with the pathfinding system, that caused the agent to get stuck in certain circumstances, such as when encountering a locked door.
* The map grid is now initialized on a just-in-time basis, because initializing an 8x53x21x79 array up front takes too long for my liking.

There's a ton of functionality in organized_agent that I still haven't translated over to observant_agent. So, uh, be patient I guess.