Fixed panic to 2 pages of monsters
If more than 10 monsters exist on the screen, when you pull up the list of all monsters on the screen, it splits into pages with 10 monsters to a page. I had foreseen this eventuality and wrote my code to deal with it sanely; but I incorrectly assumed that the message at the top (the "showing all monsters" one) would continue to be shown on page 2+. It does not, meaning my agent thought it was done with the screen, greatly confusing it when it tried to do something else and it didn't work because there was still a page of information onscreen. To fix this, I changed what the agent uses to determine if it's done looking at pages of monsters. (It now looks for the --more-- at the bottom of the screen.)