Updated Observant_Agent to work with latest NLE
An NLE update changed which action IDs correspond to which actions. This thoroughly borked the agent, getting it stuck in random menus and causing it to always panic turn 1. This commit allows the agent to react sanely to the new NLE.

Advanced_Agent and Organized_Agent are probably similarly borked, but I'm not supporting those anymore, so don't expect those to get similar updates.