Added pretty colors
Some of the text printed out by this program now displays in red, yellow, or green, so important details (fatal errors, panics, high scores, unhandled pokes, etc) catch the eye more easily. Additionally, the agent outputs more details in its end-of-run obituary. When the agent dies, the last 3 unique messages are shown (up from 1); and the desperation ratio (the proportion of in-game time spent groping for secret doors) is displayed.

A new function was added to the frontend in observant_agent.py that simply resets the agent. The agent prints its end-of-run obituary when it resets, so adding a manual reset to test_submission.py in the starter code fixes a long-standing issue where the very last run in a test submission wouldn't display its obituary.