Boulder Pushing
Agent now has basic logic to deal with boulders and will not simply always treat them as impassible. If the agent tries and fails to push a boulder, it will note the position it was at when it tried to move the boulder and not attempt to move that same boulder from that same position in the future. This should allow the agent to navigate dungeon floors more consistently.