Adding status tracking
Agent now keeps track of statuses like hunger, stunning, food poisoning, blindness, and so forth. At the end of the game, any statuses that afflicted the hero at time of death are reported in the obituary. Organized_agent had this general functionality, but only now have I added it to observant_agent. Notably, this addition brings to light how often observant_agent dies directly or indirectly of starvation – roughly 3/4 of all runs. Thus, importing food-eating behavior is next on the to-do list.