Annotated Log Report
New functionality: At the end of the run, when the agent prints the last few messages it saw before it died, monster names will be annotated with their identities. For example, the message "You hit BOB! BOB hits! BOB hits! --more--" might be edited to "You hit BOB (grid bug)! BOB hits! BOB hits! --more--" Annotations are marked cyan to denote they aren't in the original message. In addition, when a message is marked as appearing multiple times in a row, such as "You can't open anything - you have no hands! (100x)", the streak number, in this case (100x), will be labelled yellow. This is both to denote they aren't in the original message, and also to mark them as possible warning signs; if the agent does anything 100x in a row, that's probably a bad thing, I'd have to imagine.