Food Poisoning Fix
The agent will no longer eat off the floor by mistake when reaching into its inventory for chow. (This is bad because eating untrusted corpses is a great way to die of food poisoning.) In fixing this, I also added functionality for the action queue to slot 'conditional' y or n presses, that only happen if there's a yes-or-no prompt on-screen, and are skipped otherwise.

Also, I disabled "panic" as a monster name (so I can ctrl-F and only find times when the agent panics).