Basic Scavenging Behavior
Agent now has the basic framework of logic to check items on the floor and pick up ones that seem useful. Right now, the agent only uses this to snag food rations it finds, but the framework can be expanded to evaluate just about any item. The skeleton is there, we just need to put some meat on them bones. (Of course, there's a LOT of meat that needs to be applied in this manner...)

One MAJOR feature that organized_agent didn't have that observant_agent now does: NAVIGATING PILES OF MULTIPLE ITEM TYPES. It took a lot of playing around with the NLE observations, but I finally figured out how to reliably read the contents of piles from the pop-up that appears – the info is lifted straight from the raw ttychars. I'm actually *quite* happy with how the item pickup routine turned out, I'm not gonna lie.