Annotation during the fact
Annotations are now added to messages as they are logged rather than at the end of the run. This way, if a monster gets poly'd, it will only be marked as its new form after the transformation. (The way it was before, when a monster gets poly'd, the logbook would display as if the monster had always been its post-poly form.)