Added stats tracking
Agent will now track stats such as average floor reached, average turns per run, proportion of runs that die before floor 3, and so forth, and will print that information with a simple modification to test_submission.py. More information to debug with and track progress with! Also, fixed (?) a bug causing the agent to panic when it finds a square with 2 or more pages of items on it. (This is a rare occurrence, so I haven't been able to confirm the fix works...)