Added howtoforceclass.txt
This file is an instruction manual to remind me how to tinker with the NLE to always give me the class I want. Obviously, forcing a class something of a cheat – the goal is to succeed at the unmodified NLE. However, using this mod is extremely useful for testing the agent, since I can focus my attention on a single class's performance when I'm implementing behavior for something relevant to that class.

I actually had a .txt file saved on my computer to this effect. It was lost when I changed computers, so I had to rewrite it, hence why I'm adding it to the repo for safekeeping this time. It's just as well that I lost the old version of the document, I might add, because in reinventing the wheel here I actually found built-in, non-invasive ways to force a class (without poking around in the NLE library itself, which is what I did previously).