Feelin' the healin'
Agent will now use potions of healing, extra healing, and full healing when its HP gets low (either when HP is 4 or less, or when HP is 25% of max or less). This is a notable boon for monks, healers, and tourists, who are guaranteed to start with a few such potions. Wizards can sometimes also start with one, and of course any class can potentially find one in the dungeon Overall, the agent's performance is improved nicely by this change.

Fixed (?) a bug with the agent's behavior causing a panic when it encounters 2+ pages of items on one square. (This is a rare occurrence so I can't confirm the fix works yet...)